+++ synchronizer.conf
@@ -7,6 +7,10 @@ groups_search_base_list = [
         'OU=Teamcity,OU=Security Groups,DC=domainname,DC=com',
         "OU=Some Big App Comand Groups,OU=Some Big Project Teamcity Groups,OU=IT,DC=domainname,DC=com"
         ]
+custom_groups_list = [
+        "CN=SUPER_COOL_GROUP,OU=MEGA_OU,OU=Domain Groups,DC=DOMAIN,DC=SUFFIX",
+        "CN=API_NGINX_SEARCH_SUPPORT,OU=NGINX,OU=Security Groups,DC=DOMAIN,DC=SUFFIX"
+        ]
 
 [teamcity]
 server = http://teamcityServerName:teamcityPort
@@ -14,9 +18,11 @@ timeout = 10
 username = username
 password = SoMeP@$$worD312
 verify_certificate = False
-cache_file = user_state_cache.cache
-cache_ttl = 300
 
 [xml]
 path_ldap_mapping = ldap-mapping-example.xml
 #path_ldap_mapping = /opt/teamcity/datadir/config/ldap-mapping.xml
+
+[ad]
+cache_file = user_state_cache.cache
+cache_ttl = 60
